% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_addition.R
\name{module_addition}
\alias{module_addition}
\title{Title}
\usage{
module_addition(f, g)
}
\arguments{
\item{f}{polynomial}

\item{g}{polynomial}
}
\value{
the sum of the polynomials f and g
}
\description{
Title
}
\examples{
polynomial_addition(c(1, 1, 1), c(-1/2, -1))


}
