% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_create.R
\name{module_create}
\alias{module_create}
\title{Title}
\usage{
module_create(f, g)
}
\arguments{
\item{f}{polynomial}

\item{g}{polynomial}
}
\value{
vector
}
\description{
creates a 2-dimensional vector with coordinates the polynomials f and g
}
