% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_product.R
\name{module_product}
\alias{module_product}
\title{Title}
\usage{
module_product(f, g)
}
\arguments{
\item{f}{vector}

\item{g}{vector}
}
\value{
vector
}
\description{
defines a product function: Q[x]^2 x Q[x]^2 -> Q[x]
}
