% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_scalar.R
\name{module_scalar}
\alias{module_scalar}
\title{Title
scales a vector with a in Q[x]}
\usage{
module_scalar(f, a)
}
\arguments{
\item{f}{vector}

\item{a}{scalar}
}
\value{
vector
}
\description{
Title
scales a vector with a in Q[x]
}
