% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomial_addition.R
\name{polynomial_addition}
\alias{polynomial_addition}
\title{Title}
\usage{
polynomial_addition(f, g)
}
\arguments{
\item{f}{polynomial}

\item{g}{polynomial}
}
\value{
polynomial sum
}
\description{
Title
}
