% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomial_degree.R
\name{polynomial_degree}
\alias{polynomial_degree}
\title{Title}
\usage{
polynomial_degree(f)
}
\arguments{
\item{f}{polynomial}
}
\value{
degree of polynomial, 0 has degree -inf
}
\description{
Title
}
\examples{
polynomial_degree(c(1,1,1))


}
