% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomial_division.R
\name{polynomial_division}
\alias{polynomial_division}
\title{Title}
\usage{
polynomial_division(f, g)
}
\arguments{
\item{f}{polynomial}

\item{g}{polynomial}
}
\value{
polynomial
}
\description{
gives you the result after performing the division algorithm
does not give remainder
}
