% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomial_EU.R
\name{polynomial_EU}
\alias{polynomial_EU}
\title{Title}
\usage{
polynomial_EU(f, g)
}
\arguments{
\item{f}{polynomial}

\item{g}{polynomial}
}
\value{
Bezout coefficients and gcd(f,g)
}
\description{
performs the euclidean algorithm
}
\examples{

}
