% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomial_product.R
\name{polynomial_product}
\alias{polynomial_product}
\title{Title
multiplies the polynomials f and g}
\usage{
polynomial_product(f, g)
}
\arguments{
\item{f}{polynomial}

\item{g}{polynomial}
}
\value{
polynomial
}
\description{
Title
multiplies the polynomials f and g
}
