% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomial_remainder.R
\name{polynomial_remainder}
\alias{polynomial_remainder}
\title{Title}
\usage{
polynomial_remainder(f, g)
}
\arguments{
\item{f}{polynomial}

\item{g}{polynomial}
}
\value{
polynomial
}
\description{
returns the remainder after performing the division algorithm
}
