% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomial_simplify.R
\name{polynomial_simplify}
\alias{polynomial_simplify}
\title{Title}
\usage{
polynomial_simplify(f)
}
\arguments{
\item{f}{polynomial}
}
\value{
polynomial
}
\description{
removes the zero coefficients at the start
}
\examples{
polynomial_simplify(c(0,1,1,1)) gives c(1, 1, 1)

}
